+++ com.unity.render-pipelines.lightweight/Runtime/Passes/FinalBlitPass.cs
@@ -62,10 +62,13 @@ namespace UnityEngine.Rendering.LWRP
             {
                 cmd.SetGlobalTexture("_BlitTex", m_Source.Identifier());
 
+                // TODO: Final blit pass should always blit to backbuffer. The first time we do we don't need to Load contents to tile.
+                // We need to keep in the pipeline of first render pass to each render target to propertly set load/store actions.
+                // meanwhile we set to load so split screen case works.
                 SetRenderTarget(
                     cmd,
                     BuiltinRenderTextureType.CameraTarget,
-                    RenderBufferLoadAction.DontCare,
+                    RenderBufferLoadAction.Load,
                     RenderBufferStoreAction.Store,
                     ClearFlag.None,
                     Color.black,