+++ com.unity.render-pipelines.lightweight/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [5.9.0] - 2019-XX-XX
+### Fixed
+- Fixed split-screen rendering on mobile platforms.
+- Fixed rendering when using an off-screen camera that renders to a depth texture.
+- Exposed stencil render state in the renderer.
+- The default layer mask is now applied to a depth pre-pass.
+- Several improvements and fixes to the render pass UI.
 
 ## [5.8.0] - 2019-03-13
 ### Added